Move things around more for using celery
- Rely on preprint updated being called for most updates
- Use delete signal to trigger task to change update for delete
- Remove specific calls on node
